+++ src/Chamilo/Libraries/Protocol/Microsoft/Graph/Service/GroupService.php
@@ -6,6 +6,7 @@ use Chamilo\Configuration\Service\ConfigurationConsulter;
 use Chamilo\Core\User\Storage\DataClass\User;
 use Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\AzureUserNotExistsException;
 use Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GroupNotExistsException;
+use Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\UnknownAzureUserIdException;
 use Chamilo\Libraries\Protocol\Microsoft\Graph\Storage\Repository\GroupRepository;
 use GuzzleHttp\Exception\ClientException;
 use Microsoft\Graph\Model\Group;
@@ -115,7 +116,10 @@ class GroupService
      * @param string $groupName
      *
      * @return string
-     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\AzureUserNotExistsException
+     * @throws AzureUserNotExistsException
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
+     * @throws UnknownAzureUserIdException
      */
     public function createGroupByName(User $owner, $groupName)
     {
@@ -140,6 +144,9 @@ class GroupService
      *
      * @param string $groupId
      * @param string $groupName
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function updateGroupName($groupId, $groupName)
     {
@@ -153,7 +160,10 @@ class GroupService
      * @param string $groupId
      * @param \Chamilo\Core\User\Storage\DataClass\User $user
      *
-     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\AzureUserNotExistsException
+     * @throws AzureUserNotExistsException
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
+     * @throws UnknownAzureUserIdException
      */
     public function addMemberToGroup(string $groupId, User $user)
     {
@@ -170,7 +180,16 @@ class GroupService
             if ($this->isOwnerOfGroup($groupId, $user))
             {
                 $this->getGroupRepository()->subscribeMemberInGroup($groupId, $azureUserIdentifier);
-                $this->getGroupRepository()->subscribeOwnerInGroup($groupId, $azureUserIdentifier);
+
+                try
+                {
+                    $this->getGroupRepository()->subscribeOwnerInGroup($groupId, $azureUserIdentifier);
+                }
+                catch (\Exception $exception)
+                {
+                    //TODO: check if subscribe member bug is fixed. It seems so because
+                    // GRAPH api is returning already subscribed errors
+                }
             }
             else
             {
@@ -186,12 +205,21 @@ class GroupService
      * @param string $groupId
      * @param \Chamilo\Core\User\Storage\DataClass\User $user
      *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
+     * @throws AzureUserNotExistsException
      */
     public function removeMemberFromGroup($groupId, User $user)
     {
         if ($this->isMemberOfGroup($groupId, $user))
         {
             $azureUserIdentifier = $this->getAzureUserIdentifier($user);
+
+            if (empty($azureUserIdentifier))
+            {
+                throw new AzureUserNotExistsException($user);
+            }
+
             $this->getGroupRepository()->removeMemberFromGroup($groupId, $azureUserIdentifier);
         }
     }
@@ -199,6 +227,9 @@ class GroupService
     /**
      * @param string $groupId
      * @param string $azureMemberId
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function removeMemberFromGroupByAzureId(string $groupId, string $azureMemberId)
     {
@@ -212,6 +243,7 @@ class GroupService
      * @param \Chamilo\Core\User\Storage\DataClass\User $user
      *
      * @return boolean
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function isMemberOfGroup($groupId, User $user)
     {
@@ -232,6 +264,7 @@ class GroupService
      * @param string $groupId
      *
      * @return string[]
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function getGroupMembers($groupId)
     {
@@ -250,6 +283,9 @@ class GroupService
      * Removes all the members from a given group
      *
      * @param string $groupId
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function removeAllMembersFromGroup($groupId)
     {
@@ -267,7 +303,9 @@ class GroupService
      * @param string $groupId
      * @param \Chamilo\Core\User\Storage\DataClass\User $user
      *
-     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\AzureUserNotExistsException
+     * @throws AzureUserNotExistsException
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function addOwnerToGroup($groupId, User $user)
     {
@@ -280,8 +318,15 @@ class GroupService
                 throw new AzureUserNotExistsException($user);
             }
 
-            /** BUG IN MICROSOFT: THE OWNER SHOULD BE BOTH MEMBER AND OWNER  */
-            $this->getGroupRepository()->subscribeMemberInGroup($groupId, $azureUserIdentifier);
+            try
+            {
+                /** BUG IN MICROSOFT: THE OWNER SHOULD BE BOTH MEMBER AND OWNER  */
+                $this->getGroupRepository()->subscribeMemberInGroup($groupId, $azureUserIdentifier);
+            }
+            catch (\Exception $ex)
+            {
+            }
+
             $this->getGroupRepository()->subscribeOwnerInGroup($groupId, $azureUserIdentifier);
         }
     }
@@ -293,6 +338,8 @@ class GroupService
      * @param string $groupId
      * @param \Chamilo\Core\User\Storage\DataClass\User $user
      *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function removeOwnerFromGroup($groupId, User $user)
     {
@@ -306,6 +353,9 @@ class GroupService
     /**
      * @param string $groupId
      * @param string $ownerAzureId
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function removeOwnerByAzureId(string $groupId, string $ownerAzureId)
     {
@@ -319,6 +369,7 @@ class GroupService
      * @param \Chamilo\Core\User\Storage\DataClass\User $user
      *
      * @return boolean
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function isOwnerOfGroup(string $groupId, User $user)
     {
@@ -340,6 +391,8 @@ class GroupService
      * @param string $groupId
      *
      * @return string[]
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function getGroupOwners($groupId)
     {
@@ -358,6 +411,9 @@ class GroupService
      * Removes all the owners from a given group
      *
      * @param string $groupId
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function removeAllOwnersFromGroup($groupId)
     {
@@ -375,6 +431,7 @@ class GroupService
      * @param string $groupId
      *
      * @return string[]
+     * @throws \Exception
      */
     public function getGroupPlanIds($groupId)
     {
@@ -394,6 +451,7 @@ class GroupService
      * @param string $groupId
      *
      * @return string
+     * @throws \Exception
      */
     public function getDefaultGroupPlanId($groupId)
     {
@@ -414,6 +472,8 @@ class GroupService
      * @param string $planName
      *
      * @return string
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function createPlanForGroup($groupId, $planName = null)
     {
@@ -434,6 +494,8 @@ class GroupService
      * @param string $groupId
      *
      * @return string
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function getOrCreatePlanIdForGroup($groupId)
     {
@@ -453,6 +515,8 @@ class GroupService
      * @param string $groupId
      *
      * @return string
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function getGroupUrl($groupId)
     {
@@ -477,6 +541,9 @@ class GroupService
      * @param int $groupId
      * @param User[] $users
      * @param User[] | null $excludedUsersForRemoval
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function syncUsersToGroup($groupId, $users = array(), $excludedUsersForRemoval = [])
     {
@@ -541,6 +608,7 @@ class GroupService
      * @param User[] $members
      *
      * @return string[]
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function getGroupMemberAzureIdsNotInArray(string $groupId, array $members): array
     {
@@ -552,6 +620,8 @@ class GroupService
      * @param User[] $owners
      *
      * @return string[]
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function getGroupOwnerAzureIdsNotInArray(string $groupId, array $owners): array
     {
@@ -561,6 +631,9 @@ class GroupService
     /**
      * @param string $groupId
      * @param User[] $members
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function removeGroupMembersNotInArray(string $groupId, array $members)
     {
@@ -573,6 +646,9 @@ class GroupService
     /**
      * @param string $groupId
      * @param User[] $owners
+     *
+     * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function removeGroupOwnersNotInArray(string $groupId, array $owners)
     {
@@ -587,20 +663,11 @@ class GroupService
      *
      * @return Group
      * @throws GroupNotExistsException
+     * @throws \Chamilo\Libraries\Protocol\Microsoft\Graph\Exception\GraphException
      */
     public function getGroup(string $groupId): Group
     {
-        try
-        {
-            $group = $this->groupRepository->getGroup($groupId);
-        }
-        catch(ClientException $clientException)
-        {
-            if($clientException->getCode() == '404')
-            {
-                throw new GroupNotExistsException($groupId);
-            }
-        }
+        $group = $this->groupRepository->getGroup($groupId);
 
         if (!$group instanceof Group)
         {
